move lock token from message function to keeper file
